requirements.txt
Flask
requests
python-dotenv
geopy
sqlalchemy